Refactor IStore interface and implement in JsonStore
This commit introduces an ID field to IStore interface methods and applies relevant changes to the JsonStore implementation. It also introduces the Resource class and refactors the JsonStore class to further allow object identification and management. This modification is aimed at improving the efficiency and flexibility of the storage system.